+++ build-system/test-configs/forbidden-terms.js
@@ -805,10 +805,6 @@ const forbiddenTermsGlobal = {
       'test/unit/test-mustache.js',
       'test/unit/test-pass.js',
       'test/unit/test-platform.js',
-      'test/unit/test-polyfill-document-contains.js',
-      'test/unit/test-polyfill-math-sign.js',
-      'test/unit/test-polyfill-object-assign.js',
-      'test/unit/test-polyfill-object-values.js',
       'test/unit/test-pull-to-refresh.js',
       'test/unit/test-purifier.js',
       'test/unit/test-render-delaying-services.js',
