+++ test/unit/polyfills/test-document-contains.js
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-import {install} from '../../src/polyfills/document-contains';
+import {install} from '../src/polyfills/document-contains';
 
-describe('HTMLDocument.contains', () => {
+describes.sandboxed('HTMLDocument.contains', {}, () => {
   let fakeWinWithContains;
   let fakeWinWithoutContains;
   let nativeContains;
