+++ test/unit/polyfills/test-domtokenlist-toggle.js
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-import {install} from '../../src/polyfills/domtokenlist';
-import {toArray} from '../../src/core/types/array';
+import {install} from '../src/polyfills/domtokenlist';
+import {toArray} from '../src/core/types/array';
 
 describes.fakeWin(
   'DOMTokenList.toggle/add on non-IE',
